// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export type Time = bigint;
export interface LogoUpload {
    teamName: string;
    logo: ExternalBlob;
    fileSize: bigint;
    uploadTime: Time;
}
export interface SocialMediaLinks {
    instagram?: string;
    whatsapp?: string;
    discord?: string;
    youtube?: string;
}
export interface ReportMetadata {
    created: Time;
    blob: ExternalBlob;
    size: bigint;
    filename: string;
}
export interface TeamRegistration {
    id: string;
    creator?: Principal;
    teamName: string;
    members: Array<TeamMember>;
    groupNumber: bigint;
    logo: ExternalBlob;
    idPassAssigned: boolean;
    captain: CaptainInfo;
    teamId: bigint;
    groupLink?: string;
}
export interface DuplicateEntry {
    originalId: string;
    duplicatedAt: Time;
    duplicateId: string;
    reason: DuplicateReason;
}
export interface UserApprovalInfo {
    status: ApprovalStatus;
    principal: Principal;
}
export interface CaptainInfo {
    name: string;
    phone: string;
}
export interface Result {
    id: string;
    title: string;
    link?: string;
    createdAt: Time;
    description: string;
    image: ExternalBlob;
}
export interface RegistrationSummary {
    totalTeams: bigint;
    confirmedTeams: bigint;
    maxTeams?: bigint;
}
export interface TeamMember {
    playerId?: string;
    name: string;
}
export interface IdPassGroupLink {
    id: string;
    creator: Principal;
    groupNumber: bigint;
    createdAt: Time;
    groupId: bigint;
    groupLink: string;
}
export interface Announcement {
    id: string;
    title: string;
    link?: string;
    createdAt: Time;
    description: string;
    image: ExternalBlob;
}
export interface UserProfile {
    name: string;
}
export enum ApprovalStatus {
    pending = "pending",
    approved = "approved",
    rejected = "rejected"
}
export enum DuplicateReason {
    captainPhone = "captainPhone",
    teamName = "teamName"
}